refactor crud functins out of App
