+++ frontend/src/App.jsx
@@ -3,9 +3,7 @@ import { useNavigate, Outlet } from "react-router-dom";
 import { isToday, parseJSON } from "date-fns";
 import Register from "./routes/Register";
 import Login from "./routes/Login";
-import Task from "./components/Task";
 import Didit from "./components/Didit";
-import TaskForm from "./components/TaskForm";
 import ProjectForm from "./components/ProjectForm";
 import Navbar from "./components/Navbar";
 import Menu from "./components/Menu";
@@ -15,11 +13,11 @@ import taskService from "./services/tasks";
 import projectService from "./services/projects";
 import diditService from "./services/didits";
 import alertService from "./services/alerts";
-
+import CreateTask from "./components/Tasks/CreateTask";
+import ReadAndUpdateTasks from "./components/Tasks/ReadAndUpdateTasks";
 import "./App.css";
 
 function App() {
-  let navigate = useNavigate();
   const [tasks, setTasks] = useState([]);
   const [allTasks, setAllTasks] = useState([]);
   const [projects, setProjects] = useState([]);
@@ -30,9 +28,6 @@ function App() {
   const [user, setUser] = useState(null);
   const [newUser, setNewUser] = useState(false);
   const [addTask, setAddTask] = useState(false);
-  const [taskTitle, setTaskTitle] = useState("");
-  const [taskDescription, setTaskDescription] = useState("");
-  const [taskDueDate, setTaskDueDate] = useState(new Date());
   const [addProject, setAddProject] = useState(false);
   const [projectTitle, setProjectTitle] = useState("");
   const [projectId, setProjectId] = useState("");
@@ -44,16 +39,6 @@ function App() {
     setNewUser((prevVal) => !prevVal);
   };
 
-  //function to show create task form
-  const showCreateTaskForm = () => {
-    setTaskTitle("");
-    setTaskDescription("");
-    setAddTask((prevVal) => !prevVal);
-    setTasks((prevTasks) =>
-      prevTasks.map((task) => ({ ...task, isEditing: false }))
-    );
-  };
-
   //function to show create project form
   const showCreateProjectForm = () => {
     setProjectTitle("");
@@ -63,47 +48,11 @@ function App() {
     );
   };
 
-  //function to hide create task form
-  const hideCreateTaskForm = (e) => {
-    e.preventDefault();
-    setAddTask((prevVal) => !prevVal);
-    setTaskTitle("");
-    setTaskDescription("");
-    setTaskDueDate(new Date());
-  };
-
   //function to hide create project form
   const hideCreateProjectForm = (e) => {
     e.preventDefault();
     setAddProject((prevVal) => !prevVal);
     setProjectTitle("");
-  };
-
-  //function to create tasks from submission of form event
-  //posts to db with helper function and adds new task to UI
-  const handleCreateTask = async (e) => {
-    try {
-      e.preventDefault();
-      const newTask = await taskService.createTasks({
-        title: taskTitle,
-        description: taskDescription,
-        dueDate: taskDueDate,
-        isChecked: false,
-        isEditing: false,
-        project: projectId,
-      });
-      setTaskTitle("");
-      setTaskDescription("");
-      setTaskDueDate(new Date());
-      setProjectId("");
-      setTasks((prevTasks) => prevTasks.concat(newTask));
-      showCreateTaskForm();
-    } catch (error) {
-      setSystemMessage("System encountered an error");
-      setTimeout(() => {
-        setSystemMessage(null);
-      }, 3000);
-    }
   };
 
   //function to create projects from submission of form event
@@ -131,23 +80,6 @@ function App() {
 
   //function to show task form for editing task inline, populates fields with current data
   //hides the current task being edited
-  const showUpdateTaskForm = (id) => {
-    const projectId = tasks.filter((task) => task.id === id)[0].project;
-    setProjectTitle(projects.filter((task) => task.id === projectId)[0].title);
-    setAddTask(false);
-    setTaskTitle(tasks.filter((task) => task.id === id)[0].title);
-    setTaskDescription(tasks.filter((task) => task.id === id)[0].description);
-    setTasks((prevTasks) =>
-      prevTasks.map((task) =>
-        task.id === id
-          ? { ...task, isEditing: !task.isEditing }
-          : { ...task, isEditing: false }
-      )
-    );
-  };
-
-  //function to show task form for editing task inline, populates fields with current data
-  //hides the current task being edited
   const showUpdateProjectForm = (e, id) => {
     e.stopPropagation();
     setAddProject(false);
@@ -162,18 +94,6 @@ function App() {
     );
   };
 
-  //function to hide task form for editing task
-  //resets form fields on cancel
-  const hideUpdateTaskForm = () => {
-    setTaskTitle("");
-    setTaskDescription("");
-    setTaskDueDate(new Date());
-    setProjectTitle("");
-    setTasks((prevTasks) =>
-      prevTasks.map((task) => ({ ...task, isEditing: false }))
-    );
-  };
-
   //function to hide task form for editing project
   //resets form fields on cancel
   const hideUpdateProjectForm = () => {
@@ -181,74 +101,6 @@ function App() {
     setProjects((prevProjects) =>
       prevProjects.map((project) => ({ ...project, isEditing: false }))
     );
-  };
-
-  //function to update tasks from submission of form event
-  //puts to db with helper function and updates task in UI
-  const handleUpdateTask = async (e, id) => {
-    try {
-      e.preventDefault();
-      const updatedTask = tasks.filter((task) => task.id === id)[0];
-      //move these into taskservice function
-      updatedTask.title = taskTitle;
-      updatedTask.description = taskDescription;
-      updatedTask.dueDate = taskDueDate;
-      updatedTask.isEditing = false;
-      updatedTask.project = projectId;
-      await taskService.updateTasks({
-        ...updatedTask,
-      });
-
-      setTasks((prevTasks) =>
-        prevTasks.map((task) =>
-          task.id === id
-            ? {
-                ...task,
-                title: updatedTask.title,
-                description: updatedTask.description,
-                isEditing: updatedTask.isEditing,
-                project: projectId,
-              }
-            : task
-        )
-      );
-      setTaskTitle("");
-      setTaskDescription("");
-      setTaskDueDate(new Date());
-      setProjectId("");
-    } catch (error) {
-      setSystemMessage("System encountered an error");
-      setTimeout(() => {
-        setSystemMessage(null);
-      }, 3000);
-    }
-  };
-
-  //function to update if a task has been completed
-  //sets completedOn date and updates UI to show strikethrough task
-  const handleUpdateCheck = async (id) => {
-    try {
-      const currentDate = new Date();
-      const updatedTask = tasks.filter((task) => task.id === id)[0];
-      await taskService.updateTasks({
-        ...updatedTask,
-        isChecked: !updatedTask.isChecked,
-        completedOn: currentDate,
-      });
-
-      setTasks((prevTasks) =>
-        prevTasks.map((task) =>
-          task.id === id
-            ? { ...task, isChecked: !task.isChecked, completedOn: currentDate }
-            : task
-        )
-      );
-    } catch (error) {
-      setSystemMessage("System encountered an error");
-      setTimeout(() => {
-        setSystemMessage(null);
-      }, 3000);
-    }
   };
 
   //function to update project from submission of form event
@@ -281,30 +133,6 @@ function App() {
         )
       );
       setProjectTitle("");
-    } catch (error) {
-      setSystemMessage("System encountered an error");
-      setTimeout(() => {
-        setSystemMessage(null);
-      }, 3000);
-    }
-  };
-
-  //function to delete tasks from submission of form event
-  //deletes to db with helper function and updates task in UI
-  //before deleting, a didit is created from task object and posted with helper
-  const handleDeleteTask = async (id) => {
-    try {
-      const deletedTask = tasks.filter((task) => task.id === id)[0];
-      const newDidit = await diditService.createDidits(
-        { ...deletedTask },
-        user
-      );
-
-      setDidits((prevDidits) => prevDidits.concat(newDidit));
-
-      await taskService.deleteTasks(deletedTask);
-
-      setTasks((prevTasks) => prevTasks.filter((task) => task.id !== id));
     } catch (error) {
       setSystemMessage("System encountered an error");
       setTimeout(() => {
@@ -353,41 +181,6 @@ function App() {
       }
     }
   };
-
-  //map out tasks into Task Components and on edit switch out for Task Form
-  const taskElements = tasks.map((task) =>
-    task.isEditing ? (
-      <TaskForm
-        key={task.id}
-        id={task.id}
-        titleValue={taskTitle}
-        descriptionValue={taskDescription || task.description}
-        dueDate={taskDueDate}
-        onSubmit={(e) => handleUpdateTask(e, task.id)}
-        onTitleChange={(e) => setTaskTitle(e.target.value)}
-        onDescriptionChange={(e) => setTaskDescription(e.target.value)}
-        onDueDate={setTaskDueDate}
-        onClick={hideUpdateTaskForm}
-        projects={projects}
-        projectTitle={projectTitle}
-        projectId={projectId}
-        onProjectId={setProjectId}
-        submitText="save"
-      />
-    ) : (
-      <Task
-        checked={task.isChecked}
-        onCheck={handleUpdateCheck}
-        onDelete={handleDeleteTask}
-        onUpdate={showUpdateTaskForm}
-        title={task.title}
-        description={task.description}
-        dueDate={task.dueDate}
-        key={task.id}
-        id={task.id}
-      />
-    )
-  );
 
   const getDidits = async (diditTitle) => {
     if (diditTitle) {
@@ -454,9 +247,8 @@ function App() {
   //Get a user's tasks. Look into setting a timeout and "loading" screen
   useEffect(() => {
     try {
-      console.log("rendereddskfbgldskgldksr");
       const getTasks = async () => {
-        const response = await taskService.getTasks(user);
+        const response = await taskService.getTasks(user || "");
         setTasks(response);
         setAllTasks(response);
       };
@@ -548,27 +340,35 @@ function App() {
       <Button
         text="All"
         onClick={() => {
-          // setTasks(allTasks);
-          // navigate("/app/all");
+          setTasks(allTasks);
+          // navigate("/all");
         }}
       />
-      {user && taskElements}
+      {user && (
+        <ReadAndUpdateTasks
+          user={user}
+          tasks={tasks}
+          onAddTask={setAddTask}
+          onTasks={setTasks}
+          onDidits={setDidits}
+          onSystemMessage={setSystemMessage}
+          projects={projects}
+          projectTitle={projectTitle}
+          onProjectTitle={setProjectTitle}
+          projectId={projectId}
+          onProjectId={setProjectId}
+        />
+      )}
       <Outlet />
       {user && addTask ? (
-        <TaskForm
-          onSubmit={(e) => handleCreateTask(e)}
-          titleValue={taskTitle}
-          descriptionValue={taskDescription}
-          dueDate={taskDueDate}
-          onTitleChange={(e) => setTaskTitle(e.target.value)}
-          onDescriptionChange={(e) => setTaskDescription(e.target.value)}
-          onDueDate={setTaskDueDate}
-          onClick={hideCreateTaskForm}
+        <CreateTask
+          onAddTask={setAddTask}
+          onTasks={setTasks}
+          onSystemMessage={setSystemMessage}
           projects={projects}
           projectTitle={projectTitle}
           projectId={projectId}
           onProjectId={setProjectId}
-          submitText="add"
         />
       ) : (
         user && (
